Move function into own file
