+++ static/js/util/util.js
@@ -9,9 +9,7 @@ import iso3166 from 'iso-3166-2';
 import { S } from '../lib/sanctuary';
 const { Maybe, Just, Nothing } = S;
 import R from 'ramda';
-import Button from 'react-mdl/lib/Button';
 
-import SpinnerButton from '../components/SpinnerButton';
 import {
   STATUS_PASSED,
   EDUCATION_LEVELS,
@@ -460,27 +458,3 @@ export function getUserDisplayName(profile: Profile): string {
 
   return `${first} ${last}${preferred_name}`;
 }
-
-/**
- * Helper function to create dialog action buttons, with a SpinnerButton for the save button
- */
-export const dialogActions = (
-  onCancel: Function, onSave: Function, inFlight: bool, text: string='Save', saveClass: string=''
-) => ([
-  <Button
-    type='cancel'
-    key='cancel'
-    className="secondary-button cancel-button"
-    onClick={onCancel}>
-    Cancel
-  </Button>,
-  <SpinnerButton
-    component={Button}
-    spinning={inFlight}
-    type='button'
-    key='save'
-    className={`primary-button save-button ${saveClass}`}
-    onClick={onSave}>
-    {text}
-  </SpinnerButton>
-]);
