+++ static/js/util/util_test.js
@@ -29,7 +29,6 @@ import {
   classify,
   currentOrFirstIncompleteStep,
   getUserDisplayName,
-  dialogActions,
 } from '../util/util';
 import {
   EDUCATION_LEVELS,
@@ -608,28 +607,6 @@ describe('utility functions', () => {
     it('does not show preferred name when first name has same value', () => {
       profile.first_name = 'test';
       assert.equal('test doe', getUserDisplayName(profile));
-    });
-  });
-
-  describe('dialogActions', () => {
-    it('renders properly', () => {
-      let onCancel = "onCancel";
-      let onSave = "onSave";
-      let inFlight = "inFlight";
-      let newClass = "new-class";
-      let newText = "New Text";
-      let actions = dialogActions(onCancel, onSave, inFlight, newText, newClass);
-      assert.lengthOf(actions, 2);
-      let [cancelButton, saveButton] = actions;
-      assert.equal(cancelButton.props.className, "secondary-button cancel-button");
-      assert.equal(cancelButton.props.onClick, onCancel);
-      assert.equal(cancelButton.props.children, "Cancel");
-
-      assert.equal(saveButton.props.className, `primary-button save-button ${newClass}`);
-      assert.equal(saveButton.props.onClick, onSave);
-      assert.equal(saveButton.props.children, newText);
-      assert.equal(saveButton.props.spinning, inFlight);
-      assert.equal(saveButton.type.name, "SpinnerButton");
     });
   });
 });
